inherit submap merging test from loop closure handler test
