+++ algorithms/loopclosure/loop-closure-handler/CMakeLists.txt
@@ -16,7 +16,9 @@ cs_add_library(${PROJECT_NAME}
   src/loop-closure-handler.cc
   src/loop-detector-node.cc
   src/visualization/loop-closure-visualizer.cc
+  test/test_loop_closure_handling_test.cc
   ${PROTO_SRCS})
+target_link_libraries(${PROJECT_NAME} gtest)
 
 ##########
 # GTESTS #
